litd

+++ src/lnd/start_stop_lnd.sh
@@ -1,9 +1,10 @@
 function start_lnd {
-    
+    if grep -q "litd" < $ic ; then start_litd ; return 0 ; fi 
     if grep -q "lnd-" < $ic ; then
     check_rpc_bitcoin
     please_wait
     sudo systemctl start lnd.service 
+    return 0
     fi 
 
     if grep -q "lnddocker-" < $ic ; then
@@ -13,19 +14,27 @@ function start_lnd {
 
 function stop_lnd {
     please_wait
+    if grep -q "litd" < $ic ; then stop_litd ; return 0 ; fi 
+
     if grep -q "lnd-" < $ic ; then
     sudo systemctl stop lnd.service 
+    return 0
     fi
+
     if grep -q "lnddocker-" $ic ; then
     lnd_docker_stop
     fi
 }
 
 function restart_lnd {
+    if grep -q "litd" < $ic ; then restart_litd ; return 0 ; fi 
+
     if grep -q "lnd-" < $ic ; then
     check_rpc_bitcoin
     sudo systemctl restart lnd.service 
+    return 0
     fi
+
     if grep -q "lnddocker-" < $ic ; then
     lnd_docker_stop
     lnd_docker_start
